Update main.js
Refactor prevent default variable to be split between on hold and type shortcut. This allows the keydown function to determine if the current keystroke should be sent to the DOM, even if its related on hold keys are prevented.

Create edge case: if shortcut completes, remove on hold default prevention. Completed shortcuts are removed from `valid_shortcuts` which are required in order to determine if a keystroke belongs as a hold key or in a type shortcut